cupti_activity; cupti_event -> cupti_counters; float -> double in many places; python enum update

+++ source/tests/rusage_tests.cpp
@@ -49,9 +49,9 @@ static written_bytes wb;
 static auto          tot_size = nelements * sizeof(int64_t) / memory_unit.first;
 static auto          tot_rw   = nelements * sizeof(floating_t) / memory_unit.first;
 
-static const float peak_tolerance = 5 * tim::units::MiB;
-static const float curr_tolerance = 5 * tim::units::MiB;
-static const float byte_tolerance = tot_rw;  // macOS is not dependable
+static const double peak_tolerance = 5 * tim::units::MiB;
+static const double curr_tolerance = 5 * tim::units::MiB;
+static const double byte_tolerance = tot_rw;  // macOS is not dependable
 
 #define CHECK_AVAILABLE(type)                                                            \
     if(!tim::trait::is_available<type>::value)                                           \
